Move architecture to Router logic
