+++ src/telegram_bot/WebApi/Services/Bot/AuthenticatonManager.cs
@@ -1,0 +1,18 @@
+﻿using Telegram.Bot.Types;
+using TelegramForwardly.WebApi.Models.Dtos;
+
+namespace TelegramForwardly.WebApi.Services.Bot
+{
+    public static class AuthenticationManager
+    {
+        public static async Task HandlePhoneInputAsync(
+            BotUser user,
+            Message message,
+
+            CancellationToken cancellationToken
+            )
+        {
+
+        }
+    }
+}
